Implemented search path support for fortunes
the search path can be set by the MISFORTUNE_PATH env var.
a '+' prefix on any path component will make that path be searched recursively.
multiple occurrences of the same fortune db name in different places on the path will all be accepted.  This should make it really easy to create "overlays" that will allow an application to add to existing fortune databases without having to copy the original.