Improved search for fortune files in executable...
When the file can be found searching only files of the appropriate type
(based on -o, -a, etc.), then only that instance of the file is used.
Otherwise the search is repeated as if -a had been given, so that the
file will be found if at all possible.  Otherwise, barf and print usage info.